review: move LocalCatalogFactory back to polaris-core
